:art: Refactor lastz.py

+++ biolopy/lastz.py
@@ -46,43 +46,44 @@ class PairwiseAlignment:
         query_chromosomes = sorted(ensemblgenomes.rglob(patt, self._query))
         assert target_chromosomes
         assert query_chromosomes
-        chains: list[str] = []
+        chains: dict[Path, list[Path]] = {}
         with confu.ThreadPoolExecutor(max_workers=self._jobs) as executor:
             futures = [
-                executor.submit(self.lastz_chain, t, q)
+                executor.submit(self.align_chromosome_pair, t, q)
                 for t in target_chromosomes
                 for q in query_chromosomes
             ]
             (_done, _notdone) = confu.wait(futures)
             for future in futures:
-                chains.append(future.result())
-        subdirs = set(Path(x).parent for x in chains)
-        print(f"{subdirs=}")
+                chain = future.result()
+                chains.setdefault(chain.parent, []).append(chain)
         with confu.ThreadPoolExecutor(max_workers=self._jobs) as executor:
-            futures = [executor.submit(self.sub, x) for x in subdirs]
+            futures = [executor.submit(self.integrate, v) for v in chains.values()]
             (_done, _notdone) = confu.wait(futures)
             for future in futures:
-                future.result()
+                sing_maf = future.result()
+                if not sing_maf.exists():
+                    print(f"## {sing_maf} does not exist!")
 
-    def rglob(self, directory: str, pattern: str):
-        return sorted(ensemblgenomes.rglob(pattern, directory))
+    def align_chromosome_pair(self, target_2bit: Path, query_2bit: Path):
+        axtgz = self.lastz(target_2bit, query_2bit)
+        chain = self.axt_chain(target_2bit, query_2bit, axtgz)
+        return chain
 
-    def sub(self, subdir: Path):
-        pre_chain = subdir / "pre.chain.gz"
-        syntenic_net = subdir / "syntenic.net"
-        sing_maf = subdir / "sing.maf"
-        self.merge_sort_pre(subdir, pre_chain)
-        self.chain_net_syntenic(pre_chain, syntenic_net)
-        self.net_axt_maf(syntenic_net, pre_chain, sing_maf)
+    def integrate(self, chains: list[Path]):
+        pre_chain = self.merge_sort_pre(chains)
+        syntenic_net = self.chain_net_syntenic(pre_chain)
+        sing_maf = self.net_axt_maf(syntenic_net, pre_chain)
+        return sing_maf
 
-    def lastz_chain(self, target_2bit: Path, query_2bit: Path):
-        pattern = r"(?<=dna_sm.)(.+?)\.2bit$"
-        mobj = re.search(pattern, str(target_2bit))
-        assert mobj
-        target_label = mobj.group(1)
-        mobj = re.search(pattern, str(query_2bit))
-        assert mobj
-        query_label = mobj.group(1)
+    def lastz(self, target_2bit: Path, query_2bit: Path):
+        def extract_seqlabel(x: Path):
+            mobj = re.search(r"(?<=dna_sm\.)(.+)$", x.stem)
+            assert mobj
+            return mobj.group(1)
+
+        target_label = extract_seqlabel(target_2bit)
+        query_label = extract_seqlabel(query_2bit)
         subdir = self._outdir / f"{target_label}"
         if not self._dry_run:
             subdir.mkdir(0o755, exist_ok=True)
@@ -92,68 +93,77 @@ class PairwiseAlignment:
         if self._quick:
             args += " --notransition --nogapped"
         lastz = self.popen_if(not axtgz.exists(), shlex.split(args), stdout=PIPE)
-        assert lastz.stdout
-        if not self._dry_run:
+        if not axtgz.exists() and not self._dry_run:
+            assert lastz.stdout
             with gzip.open(axtgz, "wt") as fout:
                 shutil.copyfileobj(lastz.stdout, fout)
+        return axtgz
 
-        chain = re.sub(r"axt\.gz$", "chain", str(axtgz))
+    def axt_chain(self, target_2bit: Path, query_2bit: Path, axtgz: Path):
+        chain = axtgz.with_suffix("").with_suffix(".chain")
         args = (
             "axtChain -minScore=5000 -linearGap=medium"
             f" stdin {target_2bit} {query_2bit} {chain}"
         )
         gz = self.popen_if(
-            not Path(chain).exists(), ["gunzip", "-c", str(axtgz)], stdout=PIPE
+            not chain.exists(), ["gunzip", "-c", str(axtgz)], stdout=PIPE
         )
-        self.run_if(not Path(chain).exists(), shlex.split(args), stdin=gz.stdout)
+        self.run_if(not chain.exists(), shlex.split(args), stdin=gz.stdout)
         return chain
 
-    def merge_sort_pre(self, subdir: Path, outfile: Path):
-        chains = sorted([str(x) for x in subdir.glob("*.chain")])
-        args = ["chainMergeSort"] + chains
-        assert chains
-        merge = self.popen_if(not outfile.exists(), args, stdout=PIPE)
+    def merge_sort_pre(self, chains: list[Path]):
+        parent = set(x.parent for x in chains)
+        subdir = parent.pop()
+        assert not parent, "chains are in the same directory"
+        pre_chain = subdir / "pre.chain.gz"
+        args = ["chainMergeSort"] + [str(x) for x in chains]
+        merge = self.popen_if(not pre_chain.exists(), args, stdout=PIPE)
         args = f"chainPreNet stdin {self._target_sizes} {self._query_sizes} stdout"
         pre = self.popen_if(
-            not outfile.exists(), shlex.split(args), stdin=merge.stdout, stdout=PIPE
+            not pre_chain.exists(), shlex.split(args), stdin=merge.stdout, stdout=PIPE
         )
-        assert pre.stdout
-        if not outfile.exists() and not self._dry_run:
-            with gzip.open(outfile, "wt") as fout:
+        if not pre_chain.exists() and not self._dry_run:
+            assert pre.stdout
+            with gzip.open(pre_chain, "wt") as fout:
                 shutil.copyfileobj(pre.stdout, fout)
+        return pre_chain
 
-    def chain_net_syntenic(self, all_chain: Path, outfile: Path):
+    def chain_net_syntenic(self, pre_chain: Path):
+        syntenic_net = pre_chain.parent / "syntenic.net"
         gz = self.popen_if(
-            not outfile.exists(), ["gunzip", "-c", str(all_chain)], stdout=PIPE
+            not syntenic_net.exists(), ["gunzip", "-c", str(pre_chain)], stdout=PIPE
         )
         args = (
             f"chainNet stdin {self._target_sizes} {self._query_sizes} stdout /dev/null"
         )
         net = self.popen_if(
-            not outfile.exists(), shlex.split(args), stdin=gz.stdout, stdout=PIPE
+            not syntenic_net.exists(), shlex.split(args), stdin=gz.stdout, stdout=PIPE
         )
-        args = f"netSyntenic stdin {outfile}"
-        self.run_if(not outfile.exists(), shlex.split(args), stdin=net.stdout)
+        args = f"netSyntenic stdin {syntenic_net}"
+        self.run_if(not syntenic_net.exists(), shlex.split(args), stdin=net.stdout)
+        return syntenic_net
 
-    def net_axt_maf(self, syntenic_net: Path, all_chain: Path, outfile: Path):
+    def net_axt_maf(self, syntenic_net: Path, pre_chain: Path):
+        sing_maf = syntenic_net.parent / "sing.maf"
         gz = self.popen_if(
-            not outfile.exists(), ["gunzip", "-c", str(all_chain)], stdout=PIPE
+            not sing_maf.exists(), ["gunzip", "-c", str(pre_chain)], stdout=PIPE
         )
         args = f"netToAxt {syntenic_net} stdin {self._target_2bit} {self._query_2bit} stdout"
         toaxt = self.popen_if(
-            not outfile.exists(), shlex.split(args), stdin=gz.stdout, stdout=PIPE
+            not sing_maf.exists(), shlex.split(args), stdin=gz.stdout, stdout=PIPE
         )
         args = "axtSort stdin stdout"
         sort = self.popen_if(
-            not outfile.exists(), shlex.split(args), stdin=toaxt.stdout, stdout=PIPE
+            not sing_maf.exists(), shlex.split(args), stdin=toaxt.stdout, stdout=PIPE
         )
         args = (
             "axtToMaf"
             f" -tPrefix={self._target_nickname}."
             f" -qPrefix={self._query_nickname}. stdin"
-            f" {self._target_sizes} {self._query_sizes} {outfile}"
+            f" {self._target_sizes} {self._query_sizes} {sing_maf}"
         )
-        self.run_if(not outfile.exists(), shlex.split(args), stdin=sort.stdout)
+        self.run_if(not sing_maf.exists(), shlex.split(args), stdin=sort.stdout)
+        return sing_maf
 
     def popen_if(self, cond: bool, args: list[str], *pargs: Any, **kwargs: Any):
         if cond and not self._dry_run:
@@ -181,9 +191,8 @@ def main():
     parser.add_argument("-q", "--query", default=os.getenv("QUERY"))
     args = parser.parse_args()
 
-    targets = filter(None, (x.strip() for x in args.target.split(",")))
-    queries = filter(None, (x.strip() for x in args.query.split(",")))
-
+    targets = [x.strip() for x in args.target.split(",") if x]
+    queries = [x.strip() for x in args.query.split(",") if x]
     print(f"{targets=}")
     print(f"{queries=}")
     i = 0
